基本完成对预设面板的改造
